fix: target python version
Allow any valid python version to be set as the target version so that
users don't need to install a new package version each time a new Python
version is released.

Add tests for validating the target version configuration.

Clarify the meaning of the target version in the doc comment.

Fixes: #63